+++ src/main/java/WebMain.java
@@ -1,34 +1,20 @@
 import game.racing.car.event.Events;
-import game.racing.car.service.RacingGameFactory;
-import game.racing.car.service.RacingGameWebRequestHandler;
+import game.racing.car.view.impl.IndexRoute;
+import game.racing.car.view.impl.NameRoute;
+import game.racing.car.view.impl.ResultRoute;
 
-import java.util.Map;
-
-import static game.racing.car.service.RacingGameWebRequestHandler.handleNameRequest;
-import static game.racing.car.service.RacingGameWebRequestHandler.handleResultRequest;
-import static game.racing.car.view.WebConsoleView.*;
 import static spark.Spark.*;
 
 public class WebMain {
-    private static final String CAR_NAMES_REQUEST_PARAMETER = "names";
-    private static final String ROUND_COUNT_REQUEST_PARAMETER = "turn";
-
 
     public static void main(String[] args) {
         port(8080);
-        RacingGameFactory racingGameFactory = new RacingGameFactory();
 
-        get("/", (req, res) -> renderIndexPage());
+        get("/", new IndexRoute.get());
 
-        post("/name", (req, res) -> {
-            Map<String, Object> model = handleNameRequest(req.queryParams(CAR_NAMES_REQUEST_PARAMETER), racingGameFactory);
-            return renderGameStartPage(model);
-        });
+        post("/name", new NameRoute.post());
 
-        post("/result", (req, res) -> {
-            Map<String, Object> model = handleResultRequest(req.queryParams(ROUND_COUNT_REQUEST_PARAMETER), racingGameFactory);
-            return renderGameResultPage(model);
-        });
+        post("/result", new ResultRoute.post());
 
         Events.reset();
     }
